+++ pages/makeFormula.php
@@ -26,6 +26,7 @@ $mg = mysqli_fetch_assoc(mysqli_query($conn, "SELECT SUM(quantity) AS total_mg F
 $meta = mysqli_fetch_array(mysqli_query($conn, "SELECT * FROM formulasMetaData WHERE fid = '$fid'"));
 
 $settings['grp_formula'] = '0';
+$defCatClass = $settings['defCatClass'];
 
 ?><head>
   <meta charset="utf-8">
@@ -199,7 +200,7 @@ $(document).ready(function() {
                   </thead>
                   <tbody id="formula_data">
                   <?php while ($formula = mysqli_fetch_array($formula_q)) {
-					 	$ing_q = mysqli_fetch_array(mysqli_query($conn, "SELECT cas, IFRA, price, ml, profile FROM ingredients WHERE BINARY name = '".$formula['ingredient']."'"));
+					 	$ing_q = mysqli_fetch_array(mysqli_query($conn, "SELECT cas, $defCatClass, price, ml, profile FROM ingredients WHERE BINARY name = '".$formula['ingredient']."'"));
 
 						$limitIFRA = searchIFRA($ing_q['cas'],$formula['ingredient'],null,$conn);
 						$limit = explode(' - ', $limitIFRA);
@@ -233,7 +234,7 @@ $(document).ready(function() {
 								echo '<td>'.$ing_q['profile'].'</td>';
 							}
 						}
-                      echo '<td align="center" class="'.$ing_q['profile'].'">'.$ingName.'</a> '.checkIng($formula['ingredient'],$conn).'</td>';
+                      echo '<td align="center" class="'.$ing_q['profile'].'">'.$ingName.'</a> '.checkIng($formula['ingredient'],$defCatClass,$conn).'</td>';
                       echo '<td align="center">'.$formula['concentration'].'</td>';
 					  
 					  if($limit != null){
@@ -243,8 +244,8 @@ $(document).ready(function() {
 							$IFRA_WARN = 'class="alert-success"'; //VALUE IS OK
 						}
 					  }else
-					  if($ing_q['IFRA'] != null){
-					  	if($ing_q['IFRA'] < $conc_p){
+					  if($ing_q[$defCatClass] != null){
+					  	if($ing_q[$defCatClass] < $conc_p){
 							$IFRA_WARN = 'class="alert-danger"'; //VALUE IS TO HIGH AGAINST LOCAL DB
 					  	}else{
 							$IFRA_WARN = 'class="alert-success"'; //VALUE IS OK