+++ releasenotes.md
@@ -1,5 +1,6 @@
 This release includes:
 ----------------------
+You can now select the default category instead of always set to Cat4
 Changed how the % limit is been shown in ingredients page
 Versioning format changed from x.x.x to x.x
 Added release notes